Move get-root-deps
