+++ src/actions/autoinstall/get-root-deps.js
